requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
PyMySQL==1.1.1
pydantic[email]==2.9.2
python-dotenv==1.0.1
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
